refactor v2 eval control steps
